Fix: Refactor into smaller modules
Closes #28
