+++ logicTest.js_final_corrected
@@ -1,0 +1,16 @@
+// Mock necessary variables
+const suits = ['spades', 'hearts', 'diamonds', 'clubs'];
+const rankOrder = ['A','2','3','4','5','6','7','8','9','10','J','Q','K'];
+let state = { remaining: {}, aceRC: 0 };
+
+// Import modules
+const { initRemaining, pickSuit } = require('./modules/deck');
+
+// Test initialization
+initRemaining(state, rankOrder);
+console.log('Initialization successful:', Object.keys(state.remaining).length > 0);
+
+// Mock remaining cards for 'A'
+state.remaining['A'] = { spades: 8, hearts: 8, diamonds: 8, clubs: 8 };
+const pickedSuit = pickSuit('A');
+console.log(`Picked Suit for A: ${pickedSuit}`);
